+++ src/link/base_station_to_user_equipment_link.py
@@ -1,4 +1,4 @@
-from link import Link
+from link.base.link import Link
 
 
 class BaseStationToUserEquipmentLink(Link):
