+++ src/link/user_equipment_to_base_station_link.py
@@ -1,4 +1,4 @@
-from link import Link
+from link.base.link import Link
 
 class UserEquipmentToBaseStationLink(Link):
     def __init__(self, source_node, destination_node, channel):
