+++ src/transceiver/user_equipment/user_equipment.py
@@ -1,4 +1,4 @@
-from element import Element
+from transceiver.base.element import Element
